fix(ci) lock down eslint@9.30.1 to fix test:types
eslint@9.31.0 removed some of the RuleContext types (see
https://github.com/eslint/eslint/pull/19910), causing the `test:types`
CI script to fail.

Since eslint-plugin-n still needs to support 8.x versions of eslint,
the implementation that relies on RuleContexts can't be removed.

A simple solution to unblock the CI is to annotate the few places that
rely on them with @ts-ignore.
When eslint 8.x support is dropped, these annotations should be removed.
That, in turn, will expose the code that relies on eslint 8.x APIs,
which should then also be removed.

As part of this PR, the eslint dependency was bumped to ^9.31.0.
This was done in order to prevent developers from developing locally
with older 9.x versions of eslint and accidentally introducing new code
that relies of these types with properly handling the type annotations
and having it fail in the CI.

Fixes #459