fix(no-missing-import): Support data imports
Imported prefixed with 'data:' are not modules to be resolved, but the
direct data that should be imported (e.g., JavaScript code, JSON or
WASM).

This PR adds support for such imports in the no-missing-import rule.

Note:
With this change, the rule doesn't attempt to parse the data being
imported, it merely does not try to resolve it to a path (which would
fail), and thus allows those imports in the user's code.

Fixes #464